[WIP]Create region list
