+++ TrainDelayReporter_iOS/UI/Components/HamburgerMenu.swift
@@ -8,46 +8,11 @@
 
 import SwiftUI
 import QGrid
-
-struct Region: Identifiable {
-    var id = UUID()
-    var name: String
-    var colorCode: Int
-}
-
-struct RegionCell:View {
-    var region: Region
-    var body: some View {
-        ZStack {
-            VStack(alignment: .leading) {
-                HStack() {
-                    Circle()
-                        .fill(Color(hex: self.region.colorCode))
-                        .frame(width: 10, height:10) //TODO Geometryで大きさを計算
-                    Text(self.region.name)
-                }
-            }
-        }
-    }
-}
+import Combine
 
 struct HamburgerMenu: View {
-    var regionStructs:[Region] = []
-    let regions:[String: Int] = [
-        "北海道": 0x02C03C,
-        "東日本": 0x37863F,
-        "西日本": 0x0072BA,
-        "東海": 0xFF7E1C,
-        "四国": 0x00ACD1,
-        "九州": 0xF62D36,
-        "私鉄": 0x98A9D6
-    ]
-    init() {
-        //TODO ハンバーガーメニューが出現するたびに要素の順番が変わってしまうのを修正する
-        for (name, color) in regions {
-            regionStructs.append(Region(name: name as! String, colorCode: color as! Int))
-        }
-    }
+    
+    @ObservedObject var regionListInteractor = RegionListInteractor()
     
     var body: some View {
         GeometryReader { geometry in
@@ -55,7 +20,10 @@ struct HamburgerMenu: View {
                 Text("地域を選択")
                     .font(.system(size: 24))
                 Divider()
-                QGrid(self.regionStructs,
+                if self.regionListInteractor.isComplete == false {
+                    Text("地域情報を取得中")
+                }
+                QGrid(self.regionListInteractor.regionStructs,
                       columns: 1,
                       vSpacing: 20,
                       hSpacing: 20,
@@ -66,6 +34,9 @@ struct HamburgerMenu: View {
                 }
             }
             .padding(.horizontal, 20)
+            .onAppear(perform: {
+                self.regionListInteractor.setRegions()
+            })
         }
     }
 }
